Ordered journal.bib and journal-short.bib entries
The entries of the journal files have been ordered alphabetically, w.r.t. the name of the journal.
The order in journal.bib takes precedence and over the order in journal-short.bib.
Removed duplicates of SWEVO
Removed duplicate of NC as NCOMP

Corrected  entry EmmDeu2018NC in bibliography.bib, accordingly.